moving category_scroller
