+++ lib/menu_elements/category_scroller.dart
